add review functionality and update restaurant structure

+++ src/main/java/com/example/foodreview/services/RestaurantService.java
@@ -1,6 +1,10 @@
 package com.example.foodreview.services;
 
+import com.example.foodreview.dto.RestaurantRequest;
+import com.example.foodreview.dto.RestaurantResponse;
 import com.example.foodreview.entities.Restaurant;
+import com.example.foodreview.exception.ResourceNotFoundException;
+import com.example.foodreview.mapper.RestaurantMapper;
 import com.example.foodreview.repositories.RestaurantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,20 +16,38 @@ import java.util.List;
 public class RestaurantService {
 
     private final RestaurantRepository restaurantRepository;
+    private final RestaurantMapper restaurantMapper;
 
-    public List<Restaurant> findAll() {
-        return restaurantRepository.findAll();
+    public List<RestaurantResponse> findAll() {
+        return restaurantRepository.findAll().stream()
+                .map(restaurantMapper::toResponse)
+                .toList();
     }
 
-    public Restaurant findById(Long id) {
-        return restaurantRepository.findById(id).orElseThrow(()-> new RuntimeException("Restaurant not found"));
+    public RestaurantResponse findById(Long id) {
+        return restaurantRepository.findById(id)
+                .map(restaurantMapper::toResponse)
+                .orElseThrow(()-> new ResourceNotFoundException("Restaurant not found"));
     }
 
-    public Restaurant save(Restaurant restaurant) {
-        return restaurantRepository.save(restaurant);
+    public RestaurantResponse createRestaurant(RestaurantRequest request) {
+        Restaurant restaurant = restaurantMapper.toEntity(request);
+        Restaurant saved = restaurantRepository.save(restaurant);
+        return restaurantMapper.toResponse(saved);
+    }
+
+    public RestaurantResponse updateRestaurant(Long id, RestaurantRequest request) {
+        Restaurant restaurant = restaurantRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Restaurant not found"));
+        restaurantMapper.updateEntityFromRequest(request, restaurant);
+        Restaurant updated = restaurantRepository.save(restaurant);
+        return restaurantMapper.toResponse(updated);
     }
 
     public void deleteById(Long id) {
+        if(!restaurantRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Restaurant not found");
+        }
         restaurantRepository.deleteById(id);
     }
 }
